+++ CHANGELOG.md
@@ -25,7 +25,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.0] - Unreleased
+## [0.10.1] - 2025-07-02
+
+### Fixed
+
+- Avoid panicking when modifying the capacity for volatile retention memory
+  [#478](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/478).
+
+## [0.10.0] - 2025-06-20
 
 ### Added